add buff

+++ ext/braintree/brain_tree.h
@@ -446,14 +446,14 @@ public:
 
         if (useSuccessFailPolicy) {
             if (successOnAll) {
-                minimumSuccess = children.size();
+                minimumSuccess = (int)children.size();
             }
             else {
                 minimumSuccess = 1;
             }
 
             if (failOnAll) {
-                minimumFail = children.size();
+                minimumFail = (int)children.size();
             }
             else {
                 minimumFail = 1;
